Fixed imports after moved files, modify CreateDictionary and BuildIndex to parse record ids by rows
