+++ scheme/users/Trapdoor.py
@@ -1,5 +1,5 @@
 import sqlparse
-from CryptoUtils import AESSIVEncryptNonce, phiFunction, get_xor
+from utils.CryptoUtils import AESSIVEncryptNonce, phiFunction, get_xor
 from pymcl import g1, pairing
 import os
